# docs/swagger.yaml
basePath: /api
definitions:
...
  models.Notification:
    properties:
      actions:
        items:
          $ref: '#/definitions/models.NotificationAction'
        type: array
      chat_id:
        type: integer
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: integer
      is_read:
        type: boolean
      job_id:
        type: integer
      message:
        type: string
      metadata:
        additionalProperties: true
        type: object
      priority:
        $ref: '#/definitions/models.NotificationPriority'
      read_at:
        type: string
      related_user_id:
        type: integer
      title:
        type: string
      type:
        $ref: '#/definitions/models.NotificationType'
      user_id:
        type: integer
    type: object
  models.NotificationAction:
    properties:
      action:
        description: 'action type: ''view_job'', ''open_chat'', ''mark_read'', etc.'
        type: string
      label:
        type: string
      primary:
        description: is this the primary action
        type: boolean
      url:
        description: URL to navigate to
        type: string
    type: object
  models.NotificationListResponse:
    properties:
      notifications:
        items:
          $ref: '#/definitions/models.Notification'
        type: array
      pagination:
        properties:
          has_next:
            type: boolean
          has_prev:
            type: boolean
          limit:
            type: integer
          offset:
            type: integer
        type: object
      total:
        type: integer
      unread:
        type: integer
    type: object
  models.NotificationPriority:
    enum:
    - low
    - normal
    - high
    - urgent
    type: string
    x-enum-varnames:
    - NotificationPriorityLow
    - NotificationPriorityNormal
    - NotificationPriorityHigh
    - NotificationPriorityUrgent
  models.NotificationStats:
    properties:
      by_type:
        additionalProperties:
          type: integer
        type: object
      total:
        type: integer
      unread:
        type: integer
    type: object
  models.NotificationType:
    enum:
    - job_application
    - job_update
    - job_claimed
    - job_completed
    - payment
    - document_upload
    - new_job
    - review
    - message
    - system
    type: string
    x-enum-comments:
      NotificationTypeDocumentUpload: Document uploaded
      NotificationTypeJobApplication: Someone applied to your job
      NotificationTypeJobClaimed: Your job was claimed
      NotificationTypeJobCompleted: Job was completed
      NotificationTypeJobUpdate: Job status changed
      NotificationTypeMessage: New chat message
      NotificationTypeNewJob: New job matching criteria
      NotificationTypePayment: Payment related
      NotificationTypeReview: New review received
      NotificationTypeSystem: System announcements
    x-enum-descriptions:
    - Someone applied to your job
    - Job status changed
    - Your job was claimed
    - Job was completed
    - Payment related
    - Document uploaded
    - New job matching criteria
    - New review received
    - New chat message
    - System announcements
    x-enum-varnames:
    - NotificationTypeJobApplication
    - NotificationTypeJobUpdate
    - NotificationTypeJobClaimed
    - NotificationTypeJobCompleted
    - NotificationTypePayment
    - NotificationTypeDocumentUpload
    - NotificationTypeNewJob
    - NotificationTypeReview
    - NotificationTypeMessage
    - NotificationTypeSystem
...
info:
  contact: {}
  description: API для приложения MoveShare
  title: MoveShare API
  version: "1.0"
paths:
...
  /chats/{chatId}/ws:
    get:
      description: Establishes WebSocket connection for real-time chat messaging
      parameters:
      - description: Chat ID
        in: path
        name: chatId
        required: true
        type: integer
      - description: JWT token for authentication
        in: query
        name: token
        required: true
        type: string
      responses: {}
      summary: WebSocket connection for chat
      tags:
      - Chat
...
  /notifications:
    delete:
      description: Deletes all notifications for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete all notifications
      tags:
      - Notifications
    get:
      description: Returns a paginated list of notifications for the authenticated
        user
      parameters:
      - default: 20
        description: Limit number of notifications returned
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Filter by notification type
        enum:
        - job_application
        - job_update
        - payment
        - document_upload
        - new_job
        - review
        - message
        - system
        in: query
        name: type
        type: string
      - default: false
        description: Show only unread notifications
        in: query
        name: unread
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NotificationListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user notifications
      tags:
      - Notifications
...
  /notifications/stats:
    get:
      description: Returns notification statistics for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NotificationStats'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get notification statistics
      tags:
      - Notifications
  /notifications/test:
    post:
      description: Sends a test notification to the current user (for development)
      parameters:
      - default: system
        description: Notification type
        enum:
        - message
        - job
        - system
        in: query
        name: type
        type: string
      - default: Test notification
        description: Notification message
        in: query
        name: message
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Send test notification
      tags:
      - Notifications
...
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT token in format: Bearer {your_token_here}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"